pyproject.toml
[project]
name = "ai-data-quality-solution"
version = "0.1.0"
description = "AI-powered data quality management solution using LangGraph"
authors = [
    {name = "Natasha Sharma",email = "nssharma.official@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.2.0",
    "langgraph>=0.2.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.17.0",
    "pydantic>=2.0.0",
    "ipykernel (>=6.30.1,<7.0.0)",
    "streamlit>=1.28.0"
]

[tool.setuptools]
packages = ["data_quality_assistant"]
package-dir = {"" = "src"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
